moved Map to MamComponent
