+++ client/pages/trip-details.jsx
@@ -4,6 +4,7 @@ import AppContext from '../lib/app-context';
 import EditReview from '../components/edit-review';
 import Comments from '../components/comments';
 import ReviewScore from '../components/review-score';
+import MapComponent from '../components/map';
 import { Map, Marker, GoogleApiWrapper } from 'google-maps-react';
 import { geocodeByAddress, getLatLng } from 'react-places-autocomplete';
 
@@ -247,13 +248,6 @@ export class TripDetails extends React.Component {
       safetyScore
     } = this.state.trip;
 
-    const containerStyle = {
-      width: this.state.containerWidth,
-      top: '21rem',
-      height: '23rem',
-      left: '1rem',
-      border: '1px solid rgb(240,131,52)'
-    };
     return (
         <>
         <Navbar handleChange={this.handleChange}
@@ -276,23 +270,8 @@ export class TripDetails extends React.Component {
           </section>
         </article>
 
-          <div id="map-trip-details-container">
-          <Map
-            id="map-trip-details"
-            containerStyle={containerStyle}
-            google={this.props.google}
-            center={{
-              lat: this.state.mapCenter.lat,
-              lng: this.state.mapCenter.lng
-            }}
-            >
-            <Marker
-            position= {{
-              lat: this.state.mapCenter.lat,
-              lng: this.state.mapCenter.lng
-            }}
-            />
-          </Map>
+        <div id="map-trip-details-container">
+
         </div>
 
         <section id="main-photo-trip-details">
